SP2-90-Refactor codebase to update namespace usage, testing pushing without CsvController.cs and test the deployment
Removed usage of the Heatington.Utility namespace and implemented the use of Heatington.Controllers in CsvDataSource.cs and DataPoint.cs files. Also updated the namespace of SourceDataManager.cs from Heatington.Data to Heatington.SourceDataManager for better code organization.
